Drop `Passes.td` and `PassDetail.h` in main `Codegen` folder.
