+++ compiler/src/iree/compiler/Codegen/Passes.cpp
@@ -4,9 +4,6 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree-dialects/Dialect/LinalgExt/Passes/Passes.h"
-
-#include "iree/compiler/Codegen/Dialect/LoweringConfig.h"
 #include "iree/compiler/Codegen/Passes.h"
 #include "mlir/Pass/PassManager.h"
 //===---------------------------------------------------------------------===//
@@ -23,11 +20,6 @@
 namespace mlir {
 namespace iree_compiler {
 
-namespace {
-#define GEN_PASS_REGISTRATION
-#include "iree/compiler/Codegen/Passes.h.inc"
-} // namespace
-
 void registerCodegenPasses() {
   // Generated.
   registerCodegenCommonPasses();
@@ -37,59 +29,6 @@ void registerCodegenPasses() {
   registerCodegenSPIRVPasses();
   registerCodegenVMVXPasses();
   registerCodegenWGSLPasses();
-  registerPasses();
-
-  static PassPipelineRegistration<> LinalgLLVMPipeline(
-      "iree-codegen-linalg-to-llvm-pipeline",
-      "Runs the progressive lowering pipeline from Linalg to LLVM",
-      [](OpPassManager &passManager) {
-        buildLLVMCPUCodegenPassPipeline(passManager);
-      });
-
-  static PassPipelineRegistration<> LinalgNVVMPipeline(
-      "iree-codegen-linalg-to-nvvm-pipeline",
-      "Runs the progressive lowering pipeline from Linalg to NVVM",
-      [](OpPassManager &passManager) {
-        buildLLVMGPUTransformPassPipeline(passManager, false);
-      });
-
-  static PassPipelineRegistration<> LinalgROCDLPipeline(
-      "iree-codegen-linalg-to-rocdl-pipeline",
-      "Runs the progressive lowering pipeline from Linalg to ROCDL",
-      [](OpPassManager &passManager) {
-        buildLLVMGPUTransformPassPipeline(passManager, true);
-      });
-
-  static PassPipelineRegistration<> LinalgSPIRVPipeline(
-      "iree-codegen-linalg-to-spirv-pipeline",
-      "Runs the progressive lowering pipeline from linalg to SPIR-V",
-      [](OpPassManager &passManager) {
-        buildSPIRVCodegenPassPipeline(passManager, /*enableFastMath=*/false);
-      });
-
-  static PassPipelineRegistration<> LLVMCPULinkingPipeline(
-      "iree-codegen-llvmcpu-linking-pipeline",
-      "Runs the LLVMCPU HAL executable linking pipeline",
-      [](OpPassManager &passManager) {
-        buildLLVMCPULinkingPassPipeline(passManager);
-      });
-
-  static PassPipelineRegistration<> VMVXLinkingPipeline(
-      "iree-codegen-vmvx-linking-pipeline",
-      "Runs the VMVX HAL executable linking pipeline",
-      [](OpPassManager &passManager) {
-        buildVMVXLinkingPassPipeline(passManager);
-      });
-}
-
-void addCommonTargetExecutablePreprocessingPasses(OpPassManager &passManager) {
-  passManager.addNestedPass<func::FuncOp>(createTypePropagationPass());
-  passManager.addPass(createBubbleUpOrdinalOpsPass());
-  passManager.addPass(createBufferizeCopyOnlyDispatchesPass());
-  passManager.addNestedPass<func::FuncOp>(
-      IREE::LinalgExt::createDecomposeSoftmaxPass());
-  // Temporary solution to avoid large allocations due to softmax lowering.
-  passManager.addNestedPass<func::FuncOp>(createRematerializeParallelOpsPass());
 }
 
 } // namespace iree_compiler
