--- compiler/src/iree/compiler/Codegen/Passes.td
@@ -1,12 +1,0 @@
-// Copyright 2021 The IREE Authors
-//
-// Licensed under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#ifndef IREE_CODEGEN_PASSES
-#define IREE_CODEGEN_PASSES
-
-include "mlir/Pass/PassBase.td"
-
-#endif  // IREE_DIALECT_FLOW_PASSES
